refactor(power): 重构功率控制逻辑
- 修改了 PowerTask 中的功率计算方法
- 优化了 RLS 算法的使用
- 重构了 PowerUpData_t 类,增加了新的成员函数
- 调整了 CAN 发送的数据内容
- 优化了代码结构,提高了可读性和可维护性

+++ MDK-ARM/User/APP/Variable.hpp
@@ -63,7 +63,7 @@ typedef struct
 
     float Zero_cross[4];
 
-    double final_3508_Out[4];
+    float final_3508_Out[4];
     float final_6020_Out[4];
     float FF_Zero_cross[4];
